Remove cast from bool to uint as convert from bool to string supported.
BZ: 213899

The conversion from boolean to string is now supported. It arbirarilly
return "true" or "false" strings.
This patch makes use of this convert specialization within ModemProxy to
format the setParameters. The consummer of the parameters supports the
conversion from "true" or "false" as also based on convert tool.
This patch also reports an error in case of failing conversion for mic
mute and update the mapping table to match the convention of convertTo.

+++ audio_platform_state/src/ModemProxy.cpp
@@ -121,8 +121,7 @@ const std::string ModemProxy::isModemAlive(void *context) const
         return mLiteralFalseValue;
     }
     string modemAlive;
-    convertTo(static_cast<uint32_t>(proxy->mModemAudioManagerInterface->isModemAlive()),
-              modemAlive);
+    convertTo(proxy->mModemAudioManagerInterface->isModemAlive(), modemAlive);
     return modemAlive;
 }
 
@@ -134,8 +133,7 @@ const std::string ModemProxy::isModemAudioAvailable(void *context) const
         return mLiteralFalseValue;
     }
     string modemAudioAvailable;
-    convertTo(static_cast<uint32_t>(proxy->mModemAudioManagerInterface->isModemAudioAvailable()),
-              modemAudioAvailable);
+    convertTo(proxy->mModemAudioManagerInterface->isModemAudioAvailable(), modemAudioAvailable);
     return modemAudioAvailable;
 }
